Create and upload screenshots in ingestStorybook

+++ api/src/index.test.ts
@@ -18,8 +18,17 @@ describe("app", () => {
   })
 
   afterAll(async () => {
-    server.close()
-    const db = await Database()
-    await db.destroy()
+    await new Promise<void>((resolve) => {
+      server.close(() => resolve())
+    })
+
+    try {
+      const db = await Database()
+      if (db.isInitialized) {
+        await db.destroy()
+      }
+    } catch (err) {
+      // Do nothing
+    }
   })
 })
